fix: use decimal for money fields

+++ database/migrations/2023_08_08_154234_create_appointments_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->bigInteger("month");
             $table->string("patient_uid");
             $table->enum("attended", ["Yes", "No", "Pending"]);
-            $table->double("price");
-            $table->double("amount_paid");
+            $table->decimal("price", 10, 2);
+            $table->decimal("amount_paid", 10, 2);
             $table->string("service");
             $table->enum("status", ["Paid", "Unpaid"])->nullable();
             $table->string("procedure", 5000);
